+++ app/Controllers/Erp/Customfields.php
@@ -1,5 +1,9 @@
 <?php
 /**
+ * @author Bodo Desderio <rooiboktechltd@gmail.com>
+ * @copyright 2026 Rooibok Technologies. All rights reserved.
+ */
+/**
  * NOTICE OF LICENSE
  *
  * This source file is subject to the TimeHRM License
@@ -323,7 +327,7 @@ class Customfields extends BaseController {
 					'priority'  => $priority
 				];
 				$Moduleattributes = new Moduleattributes();
-				$result = $Moduleattributes->update($id,$data);	
+				$result = $Moduleattributes->where('company_id', $this->tenantCompanyId())->update($id,$data);	
 				$Return['csrf_hash'] = csrf_hash();	
 				if ($result == TRUE) {
 					$Return['result'] = lang('Main.xin_custom_field_updated_success');
@@ -369,7 +373,7 @@ class Customfields extends BaseController {
 			$id = udecode(strip_tags(trim($this->request->getPost('_token'))));
 			$Return['csrf_hash'] = csrf_hash();
 			$Moduleattributes = new Moduleattributes();
-			$result = $Moduleattributes->where('custom_field_id', $id)->delete($id);
+			$result = $Moduleattributes->where('custom_field_id', $id)->where('company_id', $this->tenantCompanyId())->delete($id);
 			if ($result == TRUE) {
 				$Return['result'] = lang('Main.xin_custom_field_deleted_success');
 			} else {
