+++ app/Controllers/Erp/Documents.php
@@ -1,5 +1,9 @@
 <?php
 /**
+ * @author Bodo Desderio <rooiboktechltd@gmail.com>
+ * @copyright 2026 Rooibok Technologies. All rights reserved.
+ */
+/**
  * NOTICE OF LICENSE
  *
  * This source file is subject to the TimeHRM License
@@ -524,7 +528,7 @@ class Documents extends BaseController {
 					];
 				}
 				$OfficialdocumentsModel = new OfficialdocumentsModel();
-				$result = $OfficialdocumentsModel->update($id,$data);	
+				$result = $OfficialdocumentsModel->where('company_id', $this->tenantCompanyId())->update($id,$data);	
 				$Return['csrf_hash'] = csrf_hash();	
 				if ($result == TRUE) {
 					$Return['result'] = lang('Success.ci_official_document_updated_msg');
@@ -589,7 +593,7 @@ class Documents extends BaseController {
 					'document_type'  => $document_type
 				];
 				$DocumentsModel = new DocumentsModel();
-				$result = $DocumentsModel->update($id,$data);	
+				$result = $DocumentsModel->where('company_id', $this->tenantCompanyId())->update($id,$data);	
 				$Return['csrf_hash'] = csrf_hash();	
 				if ($result == TRUE) {
 					$Return['result'] = lang('Success.ci_general_document_updated_msg');
@@ -654,7 +658,7 @@ class Documents extends BaseController {
 			$id = udecode(strip_tags(trim($this->request->getPost('_token'))));
 			$Return['csrf_hash'] = csrf_hash();
 			$DocumentsModel = new DocumentsModel();
-			$result = $DocumentsModel->where('document_id', $id)->delete($id);
+			$result = $DocumentsModel->where('document_id', $id)->where('company_id', $this->tenantCompanyId())->delete($id);
 			if ($result == TRUE) {
 				$Return['result'] = lang('Success.ci_general_document_deleted_msg');
 			} else {
@@ -677,7 +681,7 @@ class Documents extends BaseController {
 			$id = udecode(strip_tags(trim($this->request->getPost('_token'))));
 			$Return['csrf_hash'] = csrf_hash();
 			$OfficialdocumentsModel = new OfficialdocumentsModel();
-			$result = $OfficialdocumentsModel->where('document_id', $id)->delete($id);
+			$result = $OfficialdocumentsModel->where('document_id', $id)->where('company_id', $this->tenantCompanyId())->delete($id);
 			if ($result == TRUE) {
 				$Return['result'] = lang('Success.ci_official_document_deleted_msg');
 			} else {
