+++ app/Controllers/Erp/Holidays.php
@@ -1,5 +1,9 @@
 <?php
 /**
+ * @author Bodo Desderio <rooiboktechltd@gmail.com>
+ * @copyright 2026 Rooibok Technologies. All rights reserved.
+ */
+/**
  * NOTICE OF LICENSE
  *
  * This source file is subject to the TimeHRM License
@@ -316,7 +320,7 @@ class Holidays extends BaseController {
 					'is_publish'  => $is_publish
 				];
 				$HolidaysModel = new HolidaysModel();
-				$result = $HolidaysModel->update($id, $data);	
+				$result = $HolidaysModel->where('company_id', $this->tenantCompanyId())->update($id, $data);	
 				$Return['csrf_hash'] = csrf_hash();	
 				if ($result == TRUE) {
 					$Return['result'] = lang('Success.ci_holiday_updated_msg');
@@ -362,7 +366,7 @@ class Holidays extends BaseController {
 			$id = udecode(strip_tags(trim($this->request->getPost('_token'))));
 			$Return['csrf_hash'] = csrf_hash();
 			$HolidaysModel = new HolidaysModel();
-			$result = $HolidaysModel->where('holiday_id', $id)->delete($id);
+			$result = $HolidaysModel->where('holiday_id', $id)->where('company_id', $this->tenantCompanyId())->delete($id);
 			if ($result == TRUE) {
 				$Return['result'] = lang('Success.ci_holiday_deleted_msg');
 			} else {
