+++ app/Controllers/Erp/Leaving.php
@@ -1,5 +1,9 @@
 <?php
 /**
+ * @author Bodo Desderio <rooiboktechltd@gmail.com>
+ * @copyright 2026 Rooibok Technologies. All rights reserved.
+ */
+/**
  * NOTICE OF LICENSE
  *
  * This source file is subject to the TimeHRM License
@@ -227,7 +231,7 @@ class Leaving extends BaseController {
 					'reason'  => $reason
 				];
 				$OffModel = new OffModel();
-				$result = $OffModel->update($id,$data);
+				$result = $OffModel->where('company_id', $this->tenantCompanyId())->update($id,$data);
 				$Return['csrf_hash'] = csrf_hash();	
 				if ($result == TRUE) {
 					$Return['result'] = lang('Success.employee_update_exit_success');
@@ -369,7 +373,7 @@ class Leaving extends BaseController {
 			$id = udecode(strip_tags(trim($this->request->getPost('_token'))));
 			$Return['csrf_hash'] = csrf_hash();
 			$OffModel = new OffModel();
-			$result = $OffModel->where('exit_id', $id)->delete($id);
+			$result = $OffModel->where('exit_id', $id)->where('company_id', $this->tenantCompanyId())->delete($id);
 			if ($result == TRUE) {
 				$Return['result'] = lang('Success.employee_delete_exit_success');
 			} else {
