+++ app/Controllers/Erp/Policies.php
@@ -1,5 +1,9 @@
 <?php
 /**
+ * @author Bodo Desderio <rooiboktechltd@gmail.com>
+ * @copyright 2026 Rooibok Technologies. All rights reserved.
+ */
+/**
  * NOTICE OF LICENSE
  *
  * This source file is subject to the TimeHRM License
@@ -295,7 +299,7 @@ class Policies extends BaseController {
 					];	
 				}
 				$PolicyModel = new PolicyModel();
-				$result = $PolicyModel->update($id, $data);	
+				$result = $PolicyModel->where('company_id', $this->tenantCompanyId())->update($id, $data);	
 				$Return['csrf_hash'] = csrf_hash();	
 				if ($result == TRUE) {
 					$Return['result'] = lang('Success.ci_policy_updated_msg');
@@ -337,7 +341,7 @@ class Policies extends BaseController {
 			$id = udecode(strip_tags(trim($this->request->getPost('_token'))));
 			$Return['csrf_hash'] = csrf_hash();
 			$PolicyModel = new PolicyModel();
-			$result = $PolicyModel->where('policy_id', $id)->delete($id);
+			$result = $PolicyModel->where('policy_id', $id)->where('company_id', $this->tenantCompanyId())->delete($id);
 			if ($result == TRUE) {
 				$Return['result'] = lang('Success.ci_policy_deleted_msg');
 			} else {
